Extract individual tests into their own files.
