+++ contrib/pg_upgrade/test/integration/scenarios/ao_table.h
@@ -1,0 +1,1 @@
+void test_an_ao_table_with_data_can_be_upgraded(void **state);
