+++ contrib/pg_upgrade/test/integration/scenarios/heap_table.h
@@ -1,0 +1,1 @@
+void test_a_heap_table_with_data_can_be_upgraded(void **state);
