+++ contrib/pg_upgrade/test/integration/utilities/gpdb5-cluster.h
@@ -1,4 +1,3 @@
 
 void startGpdbFiveCluster(void);
 void stopGpdbFiveCluster(void);
-
